# openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "XFans VLink API",
        "description": "XFans VLink API 提供了一组端点，用于管理用户、付费内容和订单。",
        "version": "1.0.0",
        "contact": {
            "name": "XFans VLink 团队"
        }
    },
    "servers": [
        {
            "url": "https://xfans-vlink-api.vercel.app/api/v1",
            "description": "生产环境"
        },
        {
            "url": "http://localhost:3000/api/v1",
            "description": "开发环境"
        }
    ],
...
    "paths": {
        "/posts": {
            "get": {
                "tags": [
                    "付费内容"
                ],
                "summary": "获取所有付费内容",
                "description": "获取所有可用付费内容的列表",
                "operationId": "getAllPosts",
                "responses": {
                    "200": {
                        "description": "成功获取付费内容列表",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostListResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "付费内容"
                ],
                "summary": "创建付费内容",
                "description": "创建新付费内容",
                "operationId": "createPost",
                "requestBody": {
                    "description": "付费内容信息",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreatePostRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "付费内容创建成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/posts/{id}": {
            "get": {
                "tags": [
                    "付费内容"
                ],
                "summary": "获取单个付费内容",
                "description": "根据ID获取单个付费内容的详细信息",
                "operationId": "getPostById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "付费内容ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功获取付费内容",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "付费内容不存在",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "付费内容"
                ],
                "summary": "更新付费内容",
                "description": "更新现有付费内容的信息",
                "operationId": "updatePost",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "付费内容ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "付费内容更新信息",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdatePostRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "付费内容更新成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "付费内容不存在",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "付费内容"
                ],
                "summary": "删除付费内容",
                "description": "删除指定的付费内容",
                "operationId": "deletePost",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "付费内容ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "付费内容删除成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "付费内容不存在",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "CreatePostRequest": {
                "type": "object",
                "required": [
                    "title",
                    "image",
                    "price",
                    "tokenAddress",
                    "chainId",
                    "ownerAddress"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "description": "付费内容标题",
                        "example": "高级教程：如何使用Web3技术"
                    },
                    "image": {
                        "type": "string",
                        "format": "uri",
                        "description": "付费内容封面图片URL",
                        "example": "https://example.com/images/tutorial.jpg"
                    },
                    "price": {
                        "type": "string",
                        "description": "付费内容价格（以Wei为单位）",
                        "example": "1000000000000000000"
                    },
                    "tokenAddress": {
                        "type": "string",
                        "description": "支付代币地址",
                        "example": "0x1234567890123456789012345678901234567890"
                    },
                    "chainId": {
                        "type": "integer",
                        "description": "区块链ID",
                        "example": 1
                    },
                    "ownerAddress": {
                        "type": "string",
                        "description": "内容所有者地址",
                        "example": "0x0987654321098765432109876543210987654321"
                    }
                }
            },
            "UpdatePostRequest": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string",
                        "description": "付费内容标题",
                        "example": "高级教程：如何使用Web3技术"
                    },
                    "image": {
                        "type": "string",
                        "format": "uri",
                        "description": "付费内容封面图片URL",
                        "example": "https://example.com/images/tutorial.jpg"
                    },
                    "price": {
                        "type": "string",
                        "description": "付费内容价格（以Wei为单位）",
                        "example": "1000000000000000000"
                    },
                    "tokenAddress": {
                        "type": "string",
                        "description": "支付代币地址",
                        "example": "0x1234567890123456789012345678901234567890"
                    },
                    "chainId": {
                        "type": "integer",
                        "description": "区块链ID",
                        "example": 1
                    },
                    "ownerAddress": {
                        "type": "string",
                        "description": "内容所有者地址",
                        "example": "0x0987654321098765432109876543210987654321"
                    }
                }
            },
            "PostResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "description": "操作是否成功",
                        "example": true
                    },
                    "data": {
                        "type": "object",
                        "properties": {
                            "id": {
                                "type": "string",
                                "description": "付费内容ID",
                                "example": "123e4567-e89b-12d3-a456-426614174000"
                            },
                            "title": {
                                "type": "string",
                                "description": "付费内容标题",
                                "example": "高级教程：如何使用Web3技术"
                            },
                            "image": {
                                "type": "string",
                                "format": "uri",
                                "description": "付费内容封面图片URL",
                                "example": "https://example.com/images/tutorial.jpg"
                            },
                            "price": {
                                "type": "string",
                                "description": "付费内容价格（以Wei为单位）",
                                "example": "1000000000000000000"
                            },
                            "tokenAddress": {
                                "type": "string",
                                "description": "支付代币地址",
                                "example": "0x1234567890123456789012345678901234567890"
                            },
                            "chainId": {
                                "type": "integer",
                                "description": "区块链ID",
                                "example": 1
                            },
                            "ownerAddress": {
                                "type": "string",
                                "description": "内容所有者地址",
                                "example": "0x0987654321098765432109876543210987654321"
                            }
                        }
                    }
                }
            },
            "PostListResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "description": "操作是否成功",
                        "example": true
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "type": "object",
                            "properties": {
                                "id": {
                                    "type": "string",
                                    "description": "付费内容ID",
                                    "example": "123e4567-e89b-12d3-a456-426614174000"
                                },
                                "title": {
                                    "type": "string",
                                    "description": "付费内容标题",
                                    "example": "高级教程：如何使用Web3技术"
                                },
                                "image": {
                                    "type": "string",
                                    "format": "uri",
                                    "description": "付费内容封面图片URL",
                                    "example": "https://example.com/images/tutorial.jpg"
                                },
                                "price": {
                                    "type": "string",
                                    "description": "付费内容价格（以Wei为单位）",
                                    "example": "1000000000000000000"
                                },
                                "tokenAddress": {
                                    "type": "string",
                                    "description": "支付代币地址",
                                    "example": "0x1234567890123456789012345678901234567890"
                                },
                                "chainId": {
                                    "type": "integer",
                                    "description": "区块链ID",
                                    "example": 1
                                },
                                "ownerAddress": {
                                    "type": "string",
                                    "description": "内容所有者地址",
                                    "example": "0x0987654321098765432109876543210987654321"
                                }
                            }
                        }
                    }
                }
            },
...
            "ErrorResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "description": "操作是否成功",
                        "example": false
                    },
                    "error": {
                        "type": "object",
                        "properties": {
                            "message": {
                                "type": "string",
                                "description": "错误信息",
                                "example": "服务器内部错误"
                            },
                            "code": {
                                "type": "string",
                                "description": "错误代码",
                                "example": "INTERNAL_SERVER_ERROR"
                            }
                        }
                    }
                }
            },
            "ValidationErrorResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "description": "操作是否成功",
                        "example": false
                    },
                    "error": {
                        "type": "object",
                        "properties": {
                            "message": {
                                "type": "string",
                                "description": "错误信息",
                                "example": "验证错误"
                            },
                            "code": {
                                "type": "string",
                                "description": "错误代码",
                                "example": "VALIDATION_ERROR"
                            },
                            "details": {
                                "type": "array",
                                "items": {
                                    "type": "object",
                                    "properties": {
                                        "field": {
                                            "type": "string",
                                            "description": "字段名",
                                            "example": "title"
                                        },
                                        "message": {
                                            "type": "string",
                                            "description": "错误信息",
                                            "example": "标题不能为空"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "SuccessResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "description": "操作是否成功",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "description": "成功信息",
                        "example": "操作成功"
                    }
                }
            }
        }
    }
}